Cleanup DashboardTableViewController with Extensions
